requirements.txt
absl-py==2.0.0
adagio==0.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alabaster==0.7.16
altair==5.2.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
binaryornot==0.4.4
blinker==1.7.0
bounded-pool-executor==0.0.3
cachetools==5.3.2
certifi==2023.11.17
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.2.0
cookiecutter==2.6.0
coreforecast==0.0.16
cycler==0.12.1
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.0
django-appconf==1.0.6
django-compressor==4.5.1
django-cors-headers==4.3.1
django-plotly-dash==2.3.1
django-tailwind==3.8.0
djangorestframework==3.15.2
docutils==0.20.1
dpd-components==0.1.0
ffmpeg==1.4
filelock==3.18.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.3.2
fugue==0.9.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.41
glob2==0.7
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
holidays==0.71
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
jax==0.4.23
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.12.0
kiwisolver==1.4.5
libclang==16.0.6
lightning-utilities==0.14.3
llvmlite==0.44.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.3.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.3
mysql-connector==2.2.9
narwhals==1.37.1
networkx==3.4.2
nfoursid==1.0.1
numba==0.57.1
numpy==1.23.5
numpydoc==1.6.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
patsy==1.0.1
PDPbox==0.3.0
Pillow==10.1.0
playsound==1.2.2
plotly==5.18.0
pluggy==1.3.0
pqdm==0.2.0
propcache==0.3.1
protobuf==4.25.1
psutil==5.9.8
psycopg2-binary==2.9.10
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pydeck==0.8.1b0
pydot==2.0.0
pydub==0.25.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyod==2.0.5
pyparsing==3.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==7.4.4
python-dateutil==2.8.2
python-slugify==8.0.4
pytorch-lightning==2.5.1.post0
pytweening==1.2.0
pytz==2023.3.post1
PyYAML==6.0.2
rcssmin==1.1.2
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rjsmin==1.2.2
rpds-py==0.16.2
rsa==4.9
scikit-learn==1.6.1
scipy==1.11.4
seaborn==0.13.0
shap==0.47.2
six==1.16.0
slicer==0.0.8
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
statsforecast==2.0.1
statsmodels==0.14.4
streamlit==1.30.0
sympy==1.14.0
tabulate==0.9.0
tailwind==3.1.5b0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.7.0
torchmetrics==1.7.1
tornado==6.4
tqdm==4.66.1
triad==0.9.8
types-python-dateutil==2.9.0.20241003
typing_extensions==4.13.2
tzdata==2023.3
tzlocal==5.2
u8darts==0.35.0
urllib3==2.1.0
utilsforecast==0.2.12
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
xarray==2025.4.0
xgboost==3.0.0
yarl==1.20.0
zipp==3.17.0
gunicorn
whitenoise
python-decouple
